+++ App/.gitignore
@@ -33,9 +33,3 @@ yarn-error.*
 
 # typescript
 *.tsbuildinfo
-
-# @generated expo-cli sync-2b81b286409207a5da26e14c78851eb30d8ccbdb
-# The following patterns were generated by expo-cli
-
-expo-env.d.ts
-# @end expo-cli
