+++ App/assets/adaptive-icon.png
